requirements.txt
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
evaluate>=0.4.1
accelerate>=0.27.0
sentencepiece
scikit-learn
numpy
pandas
tqdm
matplotlib
seaborn
wandb
python-dotenv
requests
PyYAML
